Rename docs to src
