+++ src/js/Integer.js
