+++ tests/validation.test.js
@@ -1,5 +1,5 @@
-import { bracketsAreBalanced } from "../docs/js/validation"
-import { InvalidInput } from "../docs/js/Errors"
+import { bracketsAreBalanced } from "../src/js/validation"
+import { InvalidInput } from "../src/js/Errors"
 
 
 describe("Expect bracketsAreBalanced to throw", () => {
